+++ cidc_api/shared/gcloud_client.py
@@ -453,7 +453,7 @@ def revoke_download_access_from_blob_names(
 def revoke_download_access(
     user_email_list: Union[str, List[str]],
     trial_id: Optional[str],
-    upload_type: Optional[str],
+    upload_type: Optional[Union[str, List[str]]],
 ) -> None:
     """
     Revoke users' download access to all objects in a trial of a particular upload type.
@@ -523,7 +523,9 @@ def _build_trial_upload_prefixes(
             f"{trial}/{ASSAY_TO_FILEPATH[upload]}"
             for trial in trial_id
             for upload in upload_type
+            if upload in ASSAY_TO_FILEPATH
             # will not have cross-assay ie null upload in a list
+            # but could have uploads without files eg manifests
         }
     else:  # null for cross-assay
         # don't affect clinical_data