+++ CHANGELOG.md
@@ -9,6 +9,19 @@ This Changelog tracks changes to this project. The notes below include a summary
 - `fixed` for any bug fixes.
 - `security` in case of vulnerabilities.
 
+## Version `0.27.32` - 05 Jan 2023
+
+- `changed` schemas bump to fix samples/participants prefix for file permissioning
+
+## Version `0.27.31` - 03 Jan 2023
+
+- `changed` don't issue new permissions if user is disabled or not approved
+
+## Version `0.27.30` - 28 Dec 2022
+
+- `changed` permissions handling to handle each trial instead of each trial / upload_type in revoke
+- `fixed` blob prefix generation error on uploads without files
+
 ## Version `0.27.29` - 27 Dec 2022
 
 - `changed` schemas bump to add urine to manifests' type of sample
@@ -31,7 +44,7 @@ This Changelog tracks changes to this project. The notes below include a summary
 
 - `changed` schemas bump for updated ASSAY_TO_FILEPATH constant
 - `changed` get GCS blob names to deduplicate across upload_types
-- `changed` permissions handling to handle each trial instead of each trial / upload_type
+- `changed` permissions handling to handle each trial instead of each trial / upload_type in grant
 
 ## Version `0.27.25` - 01 Dec 2022